# smart_home/smart home/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能家居综合管理系统后端",
    "description": "智能家居综合管理系统 Python FastAPI 后端接口服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/homes/{home_id}/devices": {
      "post": {
        "tags": [
          "设备管理"
        ],
        "summary": "添加设备",
        "operationId": "create_device_api_homes__home_id__devices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Device Api Homes  Home Id  Devices Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "设备管理"
        ],
        "summary": "查询家庭空间下的设备列表",
        "operationId": "list_devices_api_homes__home_id__devices_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          },
          {
            "name": "room_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Devices Api Homes  Home Id  Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}": {
      "get": {
        "tags": [
          "设备管理"
        ],
        "summary": "查询设备详情",
        "operationId": "get_device_api_devices__device_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Device Api Devices  Device Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "设备管理"
        ],
        "summary": "修改设备",
        "operationId": "update_device_api_devices__device_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Device Api Devices  Device Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "设备管理"
        ],
        "summary": "删除设备",
        "operationId": "delete_device_api_devices__device_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Device Api Devices  Device Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{device_id}/control": {
      "post": {
        "tags": [
          "设备管理"
        ],
        "summary": "控制设备",
        "operationId": "control_device_api_devices__device_id__control_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceControl"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Control Device Api Devices  Device Id  Control Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices": {
      "get": {
        "tags": [
          "设备管理"
        ],
        "summary": "兼容旧接口：查询设备列表",
        "operationId": "list_devices_compat_api_devices_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "家庭空间编号",
              "title": "Home Id"
            },
            "description": "家庭空间编号"
          },
          {
            "name": "room_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "房间编号",
              "title": "Room Id"
            },
            "description": "房间编号"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Devices Compat Api Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "设备管理"
        ],
        "summary": "兼容旧接口：添加设备",
        "operationId": "create_device_compat_api_devices_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreateCompat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Device Compat Api Devices Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/homes/{home_id}/linkages": {
      "post": {
        "tags": [
          "联动规则"
        ],
        "summary": "新增联动规则",
        "operationId": "create_rule_api_homes__home_id__linkages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Rule Api Homes  Home Id  Linkages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "联动规则"
        ],
        "summary": "查询联动规则列表",
        "operationId": "list_rules_api_homes__home_id__linkages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Rules Api Homes  Home Id  Linkages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/linkages/{rule_id}": {
      "put": {
        "tags": [
          "联动规则"
        ],
        "summary": "修改联动规则",
        "operationId": "update_rule_api_linkages__rule_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkageUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Rule Api Linkages  Rule Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "联动规则"
        ],
        "summary": "删除联动规则",
        "operationId": "delete_rule_api_linkages__rule_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Rule Api Linkages  Rule Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceControl": {
        "properties": {
          "action": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Action",
            "description": "控制动作，如 switch"
          },
          "target_state": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Target State",
            "description": "目标状态，如 on/off"
          },
          "param_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Param Value",
            "description": "可选控制参数，如亮度、功率等"
          }
        },
        "type": "object",
        "required": [
          "action",
          "target_state"
        ],
        "title": "DeviceControl"
      },
      "DeviceCreate": {
        "properties": {
          "room_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Room Id",
            "description": "所属房间编号"
          },
          "device_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Device Name",
            "description": "设备名称"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type",
            "description": "设备类型"
          },
          "device_status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Status",
            "description": "设备状态",
            "default": "offline"
          },
          "is_key_device": {
            "type": "boolean",
            "title": "Is Key Device",
            "description": "是否关键安全设备",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "device_name",
          "device_type"
        ],
        "title": "DeviceCreate"
      },
      "DeviceCreateCompat": {
        "properties": {
          "room_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Room Id",
            "description": "所属房间编号"
          },
          "device_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Device Name",
            "description": "设备名称"
          },
          "device_type": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Device Type",
            "description": "设备类型"
          },
          "device_status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Status",
            "description": "设备状态",
            "default": "offline"
          },
          "is_key_device": {
            "type": "boolean",
            "title": "Is Key Device",
            "description": "是否关键安全设备",
            "default": false
          },
          "home_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Home Id",
            "description": "家庭空间编号"
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "device_name",
          "device_type",
          "home_id"
        ],
        "title": "DeviceCreateCompat"
      },
...
      "DeviceUpdate": {
        "properties": {
          "room_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Room Id",
            "description": "所属房间编号"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name",
            "description": "设备名称"
          },
          "device_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Type",
            "description": "设备类型"
          },
          "device_status": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Status",
            "description": "设备状态"
          },
          "is_key_device": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Key Device",
            "description": "是否关键安全设备"
          }
        },
        "type": "object",
        "title": "DeviceUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkageCreate": {
        "properties": {
          "rule_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Rule Name",
            "description": "规则名称"
          },
          "trigger_condition": {
            "additionalProperties": true,
            "type": "object",
            "title": "Trigger Condition",
            "description": "触发条件"
          },
          "action_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action Config",
            "description": "执行动作配置"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "是否启用",
            "default": true
          },
          "created_by": {
            "type": "integer",
            "title": "Created By",
            "description": "创建人",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "rule_name",
          "trigger_condition",
          "action_config"
        ],
        "title": "LinkageCreate"
      },
      "LinkageUpdate": {
        "properties": {
          "rule_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Name",
            "description": "规则名称"
          },
          "trigger_condition": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Condition",
            "description": "触发条件"
          },
          "action_config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Config",
            "description": "执行动作配置"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled",
            "description": "是否启用"
          }
        },
        "type": "object",
        "title": "LinkageUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}